Issue IMCMS-332: New design to super admin page: files tab - Add condition

+++ src/main/java/com/imcode/imcms/domain/service/api/DefaultTemplateService.java
@@ -94,7 +94,7 @@ public class DefaultTemplateService implements TemplateService {
     @Override
     public void delete(String templateName) {
         TemplateJPA template = templateRepository.findByName(templateName);
-        templateRepository.delete(template.getId());
+        if (template != null) templateRepository.delete(template.getId());
     }
 
     public File getTemplateDirectory() {
